requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.16.1
SQLAlchemy==1.4.52

# FastAPI
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
pydantic==1.10.15

# Data Processing
pandas==1.4.0
matplotlib==3.9.1
plotly==6.3.1
numpy==1.26.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.4.0

# Testing
pytest==8.4.2
pytest-django==4.11.1

# Additional utilities
python-dateutil==2.8.2
pytz==2022.2.1
